restructure anon_usage files
